refactor UserForm component to be stateless again
